sonar - use double instead of float
git-svn-id: http://josm.openstreetmap.de/svn/trunk@9987 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/tools/Utils.java
@@ -1470,7 +1470,7 @@ public final class Utils {
      * @param loadFactor the load factor
      * @return the initial capacity for the HashMap constructor
      */
-    public static int hashMapInitialCapacity(int nEntries, float loadFactor) {
+    public static int hashMapInitialCapacity(int nEntries, double loadFactor) {
         return (int) Math.ceil(nEntries / loadFactor);
     }
 
@@ -1489,7 +1489,7 @@ public final class Utils {
      * @return the initial capacity for the HashMap constructor
      */
     public static int hashMapInitialCapacity(int nEntries) {
-        return hashMapInitialCapacity(nEntries, 0.75f);
+        return hashMapInitialCapacity(nEntries, 0.75d);
     }
 
     /**
